// src/grpc/proto/plugin/grpc_stdio.ts
export interface StdioData {
  channel: StdioData_Channel
  data: Uint8Array
}

export enum StdioData_Channel {
  INVALID = 0,
  STDOUT = 1,
  STDERR = 2,
  UNRECOGNIZED = -1,
}
// ...
export const StdioData = {
  encode(message: StdioData, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
    if (message.channel !== 0) {
      writer.uint32(8).int32(message.channel)
    }
    if (message.data.length !== 0) {
      writer.uint32(18).bytes(message.data)
    }
    return writer
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): StdioData {
    const reader = input instanceof _m0.Reader ? input : _m0.Reader.create(input)
    let end = length === undefined ? reader.len : reader.pos + length
    const message = createBaseStdioData()
    while (reader.pos < end) {
      const tag = reader.uint32()
      switch (tag >>> 3) {
        case 1:
          if (tag !== 8) {
            break
          }

          message.channel = reader.int32() as any
          continue
        case 2:
          if (tag !== 18) {
            break
          }

          message.data = reader.bytes()
          continue
      }
      if ((tag & 7) === 4 || tag === 0) {
        break
      }
      reader.skipType(tag & 7)
    }
    return message
  },

  fromJSON(object: any): StdioData {
    return {
      channel: isSet(object.channel) ? stdioData_ChannelFromJSON(object.channel) : 0,
      data: isSet(object.data) ? bytesFromBase64(object.data) : new Uint8Array(0),
    }
  },

  toJSON(message: StdioData): unknown {
    const obj: any = {}
    if (message.channel !== 0) {
      obj.channel = stdioData_ChannelToJSON(message.channel)
    }
    if (message.data.length !== 0) {
      obj.data = base64FromBytes(message.data)
    }
    return obj
  },

  create<I extends Exact<DeepPartial<StdioData>, I>>(base?: I): StdioData {
    return StdioData.fromPartial(base ?? ({} as any))
  },
  fromPartial<I extends Exact<DeepPartial<StdioData>, I>>(object: I): StdioData {
    const message = createBaseStdioData()
    message.channel = object.channel ?? 0
    message.data = object.data ?? new Uint8Array(0)
    return message
  },
}
// ...
export interface GRPCStdioServer extends UntypedServiceImplementation {
  /**
   * StreamStdio returns a stream that contains all the stdout/stderr.
   * This RPC endpoint must only be called ONCE. Once stdio data is consumed
   * it is not sent again.
   *
   * Callers should connect early to prevent blocking on the plugin process.
   */
  streamStdio: handleServerStreamingCall<Empty, StdioData>
}
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined

export type DeepPartial<T> = T extends Builtin
  ? T
  : T extends globalThis.Array<infer U>
    ? globalThis.Array<DeepPartial<U>>
    : T extends ReadonlyArray<infer U>
      ? ReadonlyArray<DeepPartial<U>>
      : T extends {}
        ? { [K in keyof T]?: DeepPartial<T[K]> }
        : Partial<T>

type KeysOfUnion<T> = T extends T ? keyof T : never
export type Exact<P, I extends P> = P extends Builtin
  ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & {
      [K in Exclude<keyof I, KeysOfUnion<P>>]: never
    }