// src/grpc/proto/proto/main.ts
export interface RunCommandRequest {
  additionalInfo: AdditionalInfo | undefined
  args: string[]
  coreCliHelperId: number
}

export interface RunCommandResponse {}

export interface AdditionalInfo {
  isTerminal: IsTerminal | undefined
  terminalDimensions: TerminalDimensions | undefined
}

export interface IsTerminal {
  stdin: boolean
  stdout: boolean
  stderr: boolean
}

export interface TerminalDimensions {
  width: number
  height: number
}
// ...
export const RunCommandRequest = {
  encode(
    message: RunCommandRequest,
    writer: _m0.Writer = _m0.Writer.create(),
  ): _m0.Writer {
    if (message.additionalInfo !== undefined) {
      AdditionalInfo.encode(message.additionalInfo, writer.uint32(10).fork()).ldelim()
    }
    for (const v of message.args) {
      writer.uint32(18).string(v!)
    }
    if (message.coreCliHelperId !== 0) {
      writer.uint32(24).uint32(message.coreCliHelperId)
    }
    return writer
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): RunCommandRequest {
    const reader = input instanceof _m0.Reader ? input : _m0.Reader.create(input)
    let end = length === undefined ? reader.len : reader.pos + length
    const message = createBaseRunCommandRequest()
    while (reader.pos < end) {
      const tag = reader.uint32()
      switch (tag >>> 3) {
        case 1:
          if (tag !== 10) {
            break
          }

          message.additionalInfo = AdditionalInfo.decode(reader, reader.uint32())
          continue
        case 2:
          if (tag !== 18) {
            break
          }

          message.args.push(reader.string())
          continue
        case 3:
          if (tag !== 24) {
            break
          }

          message.coreCliHelperId = reader.uint32()
          continue
      }
      if ((tag & 7) === 4 || tag === 0) {
        break
      }
      reader.skipType(tag & 7)
    }
    return message
  },

  fromJSON(object: any): RunCommandRequest {
    return {
      additionalInfo: isSet(object.additionalInfo)
        ? AdditionalInfo.fromJSON(object.additionalInfo)
        : undefined,
      args: globalThis.Array.isArray(object?.args)
        ? object.args.map((e: any) => globalThis.String(e))
        : [],
      coreCliHelperId: isSet(object.coreCliHelperId)
        ? globalThis.Number(object.coreCliHelperId)
        : 0,
    }
  },

  toJSON(message: RunCommandRequest): unknown {
    const obj: any = {}
    if (message.additionalInfo !== undefined) {
      obj.additionalInfo = AdditionalInfo.toJSON(message.additionalInfo)
    }
    if (message.args?.length) {
      obj.args = message.args
    }
    if (message.coreCliHelperId !== 0) {
      obj.coreCliHelperId = Math.round(message.coreCliHelperId)
    }
    return obj
  },

  create<I extends Exact<DeepPartial<RunCommandRequest>, I>>(
    base?: I,
  ): RunCommandRequest {
    return RunCommandRequest.fromPartial(base ?? ({} as any))
  },
  fromPartial<I extends Exact<DeepPartial<RunCommandRequest>, I>>(
    object: I,
  ): RunCommandRequest {
    const message = createBaseRunCommandRequest()
    message.additionalInfo =
      object.additionalInfo !== undefined && object.additionalInfo !== null
        ? AdditionalInfo.fromPartial(object.additionalInfo)
        : undefined
    message.args = object.args?.map(e => e) || []
    message.coreCliHelperId = object.coreCliHelperId ?? 0
    return message
  },
}

function createBaseRunCommandResponse(): RunCommandResponse {
  return {}
}

export const RunCommandResponse = {
  encode(_: RunCommandResponse, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
    return writer
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): RunCommandResponse {
    const reader = input instanceof _m0.Reader ? input : _m0.Reader.create(input)
    let end = length === undefined ? reader.len : reader.pos + length
    const message = createBaseRunCommandResponse()
    while (reader.pos < end) {
      const tag = reader.uint32()
      switch (tag >>> 3) {
      }
      if ((tag & 7) === 4 || tag === 0) {
        break
      }
      reader.skipType(tag & 7)
    }
    return message
  },

  fromJSON(_: any): RunCommandResponse {
    return {}
  },

  toJSON(_: RunCommandResponse): unknown {
    const obj: any = {}
    return obj
  },

  create<I extends Exact<DeepPartial<RunCommandResponse>, I>>(
    base?: I,
  ): RunCommandResponse {
    return RunCommandResponse.fromPartial(base ?? ({} as any))
  },
  fromPartial<I extends Exact<DeepPartial<RunCommandResponse>, I>>(
    _: I,
  ): RunCommandResponse {
    const message = createBaseRunCommandResponse()
    return message
  },
}

function createBaseAdditionalInfo(): AdditionalInfo {
  return { isTerminal: undefined, terminalDimensions: undefined }
}

export const AdditionalInfo = {
  encode(message: AdditionalInfo, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
    if (message.isTerminal !== undefined) {
      IsTerminal.encode(message.isTerminal, writer.uint32(10).fork()).ldelim()
    }
    if (message.terminalDimensions !== undefined) {
      TerminalDimensions.encode(
        message.terminalDimensions,
        writer.uint32(18).fork(),
      ).ldelim()
    }
    return writer
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): AdditionalInfo {
    const reader = input instanceof _m0.Reader ? input : _m0.Reader.create(input)
    let end = length === undefined ? reader.len : reader.pos + length
    const message = createBaseAdditionalInfo()
    while (reader.pos < end) {
      const tag = reader.uint32()
      switch (tag >>> 3) {
        case 1:
          if (tag !== 10) {
            break
          }

          message.isTerminal = IsTerminal.decode(reader, reader.uint32())
          continue
        case 2:
          if (tag !== 18) {
            break
          }

          message.terminalDimensions = TerminalDimensions.decode(reader, reader.uint32())
          continue
      }
      if ((tag & 7) === 4 || tag === 0) {
        break
      }
      reader.skipType(tag & 7)
    }
    return message
  },

  fromJSON(object: any): AdditionalInfo {
    return {
      isTerminal: isSet(object.isTerminal)
        ? IsTerminal.fromJSON(object.isTerminal)
        : undefined,
      terminalDimensions: isSet(object.terminalDimensions)
        ? TerminalDimensions.fromJSON(object.terminalDimensions)
        : undefined,
    }
  },

  toJSON(message: AdditionalInfo): unknown {
    const obj: any = {}
    if (message.isTerminal !== undefined) {
      obj.isTerminal = IsTerminal.toJSON(message.isTerminal)
    }
    if (message.terminalDimensions !== undefined) {
      obj.terminalDimensions = TerminalDimensions.toJSON(message.terminalDimensions)
    }
    return obj
  },

  create<I extends Exact<DeepPartial<AdditionalInfo>, I>>(base?: I): AdditionalInfo {
    return AdditionalInfo.fromPartial(base ?? ({} as any))
  },
  fromPartial<I extends Exact<DeepPartial<AdditionalInfo>, I>>(
    object: I,
  ): AdditionalInfo {
    const message = createBaseAdditionalInfo()
    message.isTerminal =
      object.isTerminal !== undefined && object.isTerminal !== null
        ? IsTerminal.fromPartial(object.isTerminal)
        : undefined
    message.terminalDimensions =
      object.terminalDimensions !== undefined && object.terminalDimensions !== null
        ? TerminalDimensions.fromPartial(object.terminalDimensions)
        : undefined
    return message
  },
}

function createBaseIsTerminal(): IsTerminal {
  return { stdin: false, stdout: false, stderr: false }
}

export const IsTerminal = {
  encode(message: IsTerminal, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
    if (message.stdin !== false) {
      writer.uint32(8).bool(message.stdin)
    }
    if (message.stdout !== false) {
      writer.uint32(16).bool(message.stdout)
    }
    if (message.stderr !== false) {
      writer.uint32(24).bool(message.stderr)
    }
    return writer
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): IsTerminal {
    const reader = input instanceof _m0.Reader ? input : _m0.Reader.create(input)
    let end = length === undefined ? reader.len : reader.pos + length
    const message = createBaseIsTerminal()
    while (reader.pos < end) {
      const tag = reader.uint32()
      switch (tag >>> 3) {
        case 1:
          if (tag !== 8) {
            break
          }

          message.stdin = reader.bool()
          continue
        case 2:
          if (tag !== 16) {
            break
          }

          message.stdout = reader.bool()
          continue
        case 3:
          if (tag !== 24) {
            break
          }

          message.stderr = reader.bool()
          continue
      }
      if ((tag & 7) === 4 || tag === 0) {
        break
      }
      reader.skipType(tag & 7)
    }
    return message
  },

  fromJSON(object: any): IsTerminal {
    return {
      stdin: isSet(object.stdin) ? globalThis.Boolean(object.stdin) : false,
      stdout: isSet(object.stdout) ? globalThis.Boolean(object.stdout) : false,
      stderr: isSet(object.stderr) ? globalThis.Boolean(object.stderr) : false,
    }
  },

  toJSON(message: IsTerminal): unknown {
    const obj: any = {}
    if (message.stdin !== false) {
      obj.stdin = message.stdin
    }
    if (message.stdout !== false) {
      obj.stdout = message.stdout
    }
    if (message.stderr !== false) {
      obj.stderr = message.stderr
    }
    return obj
  },

  create<I extends Exact<DeepPartial<IsTerminal>, I>>(base?: I): IsTerminal {
    return IsTerminal.fromPartial(base ?? ({} as any))
  },
  fromPartial<I extends Exact<DeepPartial<IsTerminal>, I>>(object: I): IsTerminal {
    const message = createBaseIsTerminal()
    message.stdin = object.stdin ?? false
    message.stdout = object.stdout ?? false
    message.stderr = object.stderr ?? false
    return message
  },
}

function createBaseTerminalDimensions(): TerminalDimensions {
  return { width: 0, height: 0 }
}

export const TerminalDimensions = {
  encode(
    message: TerminalDimensions,
    writer: _m0.Writer = _m0.Writer.create(),
  ): _m0.Writer {
    if (message.width !== 0) {
      writer.uint32(8).uint32(message.width)
    }
    if (message.height !== 0) {
      writer.uint32(16).uint32(message.height)
    }
    return writer
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): TerminalDimensions {
    const reader = input instanceof _m0.Reader ? input : _m0.Reader.create(input)
    let end = length === undefined ? reader.len : reader.pos + length
    const message = createBaseTerminalDimensions()
    while (reader.pos < end) {
      const tag = reader.uint32()
      switch (tag >>> 3) {
        case 1:
          if (tag !== 8) {
            break
          }

          message.width = reader.uint32()
          continue
        case 2:
          if (tag !== 16) {
            break
          }

          message.height = reader.uint32()
          continue
      }
      if ((tag & 7) === 4 || tag === 0) {
        break
      }
      reader.skipType(tag & 7)
    }
    return message
  },

  fromJSON(object: any): TerminalDimensions {
    return {
      width: isSet(object.width) ? globalThis.Number(object.width) : 0,
      height: isSet(object.height) ? globalThis.Number(object.height) : 0,
    }
  },

  toJSON(message: TerminalDimensions): unknown {
    const obj: any = {}
    if (message.width !== 0) {
      obj.width = Math.round(message.width)
    }
    if (message.height !== 0) {
      obj.height = Math.round(message.height)
    }
    return obj
  },

  create<I extends Exact<DeepPartial<TerminalDimensions>, I>>(
    base?: I,
  ): TerminalDimensions {
    return TerminalDimensions.fromPartial(base ?? ({} as any))
  },
  fromPartial<I extends Exact<DeepPartial<TerminalDimensions>, I>>(
    object: I,
  ): TerminalDimensions {
    const message = createBaseTerminalDimensions()
    message.width = object.width ?? 0
    message.height = object.height ?? 0
    return message
  },
}
// ...
export interface MainServer extends UntypedServiceImplementation {
  runCommand: handleUnaryCall<RunCommandRequest, RunCommandResponse>
}
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined

export type DeepPartial<T> = T extends Builtin
  ? T
  : T extends globalThis.Array<infer U>
    ? globalThis.Array<DeepPartial<U>>
    : T extends ReadonlyArray<infer U>
      ? ReadonlyArray<DeepPartial<U>>
      : T extends {}
        ? { [K in keyof T]?: DeepPartial<T[K]> }
        : Partial<T>

type KeysOfUnion<T> = T extends T ? keyof T : never
export type Exact<P, I extends P> = P extends Builtin
  ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & {
      [K in Exclude<keyof I, KeysOfUnion<P>>]: never
    }